+++ lib/screen/base_screens/base_screen.dart
@@ -1,0 +1,9 @@
+class BaseScreen {
+  BaseScreen({
+    this.previousPage,
+  });
+
+  /// The previous page of the screen.
+  /// If the previous page is null, open the home screen instead of closing app.
+  final BaseScreen? previousPage;
+}