+++ lib/screen/base_screens/scaffold_screen.dart
@@ -1,20 +1,23 @@
 import 'package:flutter/material.dart';
 
+import 'base_screen.dart';
+
 /// BaseScreen is a base class to create a screen.
-class BaseScreen {
-  const BaseScreen({
-    required this.screen,
+class ScaffoldScreen extends BaseScreen {
+  ScaffoldScreen({
+    required this.body,
     this.appBar,
     this.footer,
     this.endDrawer,
-    this.previousPage,
-  });
+    BaseScreen? previousPage,
+    this.activeDrawer = true,
+  }) : super(previousPage: previousPage);
 
   /// The appbar of the screen which is displayed on top of the screen.
   final PreferredSizeWidget? appBar;
 
   /// The main content of the screen.
-  final Widget screen;
+  final Widget body;
 
   /// The footer of the screen which is displayed at the bottom of the screen.
   final StatelessWidget? footer;
@@ -23,7 +26,6 @@ class BaseScreen {
   /// You can drag the endDrawer to open it.
   final StatelessWidget? endDrawer;
 
-  /// The previous page of the screen.
-  /// If the previous page is null, open the home screen instead of closing app.
-  final BaseScreen? previousPage;
+  /// Whether can open the drawer in this screen
+  final bool activeDrawer;
 }