+++ lib/screen/base_screens/widget_screen.dart
@@ -1,0 +1,9 @@
+import 'package:easy_app/screen/base_screens/base_screen.dart';
+import 'package:flutter/material.dart';
+
+class WidgetScreen extends BaseScreen {
+  WidgetScreen({
+    required this.child,
+  });
+  final Widget child;
+}