+++ lib/screen/exceptions/unknown_base_screen_exception.dart
@@ -1,0 +1,4 @@
+class UnknownBaseScreenException implements Exception {
+  @override
+  String toString() => "UnknownBaseScreenException";
+}