+++ lib/screen/page_manager.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'base_screen.dart';
+import 'base_screens/base_screen.dart';
 
 /// PageManager navigates between screens.
 class PageManager {