+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.14
 
 - Made `route` property of `ConfigFile` modifiable.
+- Fixed occurrences of type mismatch exception in `ConfigFile`
 
 ## 0.1.13